svn4946
core:
@par unsigned <-> size_t warningu, chyba ve vypisu kdyby textury zabiraly >4GB

+++ include/Lightsprint/RRDebug.h
@@ -155,7 +155,7 @@ namespace rr
 		static RRReporter* createWindowedReporter(class RRDynamicSolver*& solver, const char* caption = NULL, bool closeWhenDone = false);
 
 		//! Helper, converts number of bytes to human readable string, e.g. 12345678 to "12 MB".
-		static const char* bytesToString(unsigned bytes);
+		static const char* bytesToString(size_t bytes);
 	};
 
 	//////////////////////////////////////////////////////////////////////////////
